==> Crawling/5. Deduplication.ipynb <==
import json
import pandas as pd
import numpy as np

with open('Final_recipe_data.json', encoding="utf-8") as f:
    data = json.load(f)

## 1. 중복 레시피 제거

## 기존 레시피 개수 확인
print("기본 레시피 개수: ", len(data['data']))

title_list = []
del_idx_list = []
idx = 0
for datas in data['data']:
    if(datas["title"] not in title_list):
        title_list.append(datas["title"])
    else:
        del_idx_list.append(idx)
    idx = idx + 1
print("중복을 제외한 레시피 개수: ", len(title_list))
print("삭제할 레시피 개수: ",len(del_idx_list))

cnt = 0
for idx in del_idx_list:
    idx = idx - cnt
    del(data['data'][idx])
    # 앞에 재료가 하나 삭제 되면 인덱스가 하나씩 밀리게 되므로 처리
    cnt = cnt + 1 

print("중복이 제거 되었는지 확인", len(data['data']))

# 2. 중복 재료 제거
before = 0
after = 0
for datas in data['data']:
    before = before + len(datas['ingredients'])
    datas['ingredients'] = set(datas['ingredients'])
    datas['ingredients'] = list(datas['ingredients'])
    after = after + len(datas['ingredients'])

print("중복제거전 재료의 개수: ", before, "/ 중복제거 후:", after)

with open('Deduplicated_Filtered_recipe_data_Final.json', "w", encoding='utf-8') as outfile:      
                    json.dump(data, outfile, indent="\t", ensure_ascii=False)

과일                Fruit => F001 ~ 200: 이윤석, F201 ~ F400: 송지민, F401 ~ F600 : 전우진, F601~F800: 홍진우 

정육/계란           Meat => M

냉장/냉동/간편식    Retort => R   (밀키트 포함)

델리/샐러드         Delhi => D

통조림/즉석밥/면    Can => C

수산/건어물         Seafood => S

김치/반찬           Kinchi => K

쌀/잡곡             Grain => G

베이커리            Bakery => B 

장/양념/소스        Spices => Sp

간식/떡/빙과        Snack => Sn

커피/음료           Beverage  => Be

우유/유제품         Milk  => Mi

채소                Vegetable => V

with open('Deduplicated_Filtered_recipe_data_Final.json', encoding="utf-8") as f:
    data = json.load(f)

# Json파일에서 각 레시피의 재료 리스트 뽑아오기
ingredients_list = []
for i in range(len(data['data'])):
    ingredients_list.append(data['data'][i]['ingredients'])

# 뽑아온 재료 리스트를 하나의 리스트에 통합
ingredients = []
for i in ingredients_list:
    for j in i:
        ingredients.append(j)

# 리스트를 데이터 프레임으로 만들어 동일한 재료가 들어간 횟수를 카운트
sum = pd.DataFrame(ingredients, columns=["ingredient"])
df_value_counts = sum.value_counts(ascending=False)

# 엑셀 파일로 저장
sum.value_counts(ascending=False).to_excel("Deduplicated_Filtered_Ingredients_List_up_by_value_counts.xlsx")
